price tweaking

+++ app/src/main/java/com/example/cryptoappreal/DataCurrency.java
@@ -8,6 +8,6 @@ class DataCurrency {
 
     public String currencyID;
 
-    public int currencyPrice;
+    public double currencyPrice;
 
 }
